Extract PyPI search to shared utility
Move duplicated PyPI search code from PipPackageManager and
UvPackageManager into a shared pypiSearch.ts module.
